pyproject.toml
[project]
name = "payman-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.20",
    "langchain-google-genai>=2.1.1",
    "langchain-openai>=0.3.9",
    "langchain[openai]>=0.3.21",
    "langgraph>=0.3.18",
    "matplotlib>=3.10.1",
    "numexpr>=2.10.2",
    "openai>=1.68.2",
    "paymanai>=2.7.0",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "streamlit>=1.43.2",
    "tavily-python>=0.5.1",
    "wikipedia>=1.4.0",
]
